Created a sidenav toolbar with drag functionality of elements
Elements can now be dragged from the toolbar onto the drawing canvas, the canvas responds to the drop event by drawing a shape at the designated coordinates. Also added the shape to a set() so it can hold multiple elemnts, allowing us to develop the 'hide' functionality and the addText functionality. The elements have an attached 'onclik' event listener (currently they change colours for testing purposes).