Saving and Loading the elements
Canvas elements / sets can now be saved and re-loaded back to the canvas. The drag function can be re-attached upon re-loading the sets, connections have to be re-drawn, as old connections are drawn as static images. Many attributes of the elements are lost from the conversion toJSON, therefore the functionalities don't work anymore. Need to figure out how to solve this!